pyproject.toml
[project]
name = "control-equipos-erp"
version = "1.3.0"
description = "Plataforma integral para la gestión y control de inventario de equipos."
requires-python = ">=3.11"
dependencies = [
    # FastAPI y Servidor
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.41.0",
    "python-multipart>=0.0.20",
    # Base de Datos
    "SQLAlchemy>=2.0.48",
    "alembic>=1.18.4",
    "psycopg[binary]>=3.2.0",
    # Seguridad
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Serialización y Validación
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "email-validator>=2.2.0",
    # Utilidades y Archivos
    "python-dotenv>=1.0.1",
    "fpdf2>=2.8.0",
    "aiofiles>=24.1.0",
    "colorama>=0.4.6",
    # Análisis de Datos (Excel)
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    # Procesos en Segundo Plano
    "celery[redis]>=5.6.2",
    # Cloud Storage (S3 / MinIO)
    "boto3>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.0",
    "boto3-stubs[s3]>=1.35.0",
]

[dependency-groups]
dev = [
    "boto3-stubs[s3]>=1.42.64",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]
asyncio_mode = "auto"
